requirements.txt
# requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
pypdf
python-docx
chromadb
fastapi
uvicorn
python-dotenv